refactor: move auth_bloc to authentication/
