+++ lib/view/authentication/authentication.dart
@@ -1,2 +1,3 @@
-export 'package:simple_notes_app/view/authentication/sign_in/bloc/sign_in_bloc.dart';
+export 'package:simple_notes_app/view/authentication/sign_in/bloc/auth/auth_bloc.dart';
+export 'package:simple_notes_app/view/authentication/sign_in/bloc/sign_in/sign_in_bloc.dart';
 export 'package:simple_notes_app/view/authentication/sign_in/sign_in_screen.dart';
