+++ lib/view/authentication/sign_in/bloc/auth/auth_event.dart
